requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
boto3==1.35.68
dependency-injector==4.43.0
pydantic==2.10.1
python-multipart==0.0.17
python-dotenv==1.0.1
